--- a/lib/core/tree.js
+++ b/lib/core/tree.js
@@ -83,7 +83,10 @@ class TstNode {
     while (node !== null && index < keylength) {
       let code = key[index]
       // A-Z
-      if (code >= 0x41 && code <= 0x5a) {
+      // First check if it is bigger than 0x5a.
+      // Lowercase letters have higher char codes than uppercase ones.
+      // Also we assume that headers will mostly contain lowercase characters.
+      if (code <= 0x5a && code >= 0x41) {
         // Lowercase for uppercase.
         code |= 32
       }
@@ -109,7 +112,7 @@ class TernarySearchTree {
 
   /**
    * @param {Uint8Array} key
-   * @param {any} value
+   * @param {void}
    * */
   insert (key, value) {
     if (this.node === null) {
@@ -121,6 +124,7 @@ class TernarySearchTree {
 
   /**
    * @param {Uint8Array} key
+   * @return {TstNode | null}
    */
   lookup (key) {
     return this.node?.search(key)?.value ?? null