--- a/lib/core/tree.js
+++ b/lib/core/tree.js
@@ -112,7 +112,7 @@ class TernarySearchTree {
 
   /**
    * @param {Uint8Array} key
-   * @param {void}
+   * @param {any} value
    * */
   insert (key, value) {
     if (this.node === null) {