--- a/lib/core/tree.js
+++ b/lib/core/tree.js
@@ -124,7 +124,7 @@ class TernarySearchTree {
 
   /**
    * @param {Uint8Array} key
-   * @return {TstNode | null}
+   * @return {any}
    */
   lookup (key) {
     return this.node?.search(key)?.value ?? null