before:

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/TernarySearchTree.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark           time (avg)             (min … max)       p75       p99      p999
------------------------------------------------------ -----------------------------
• tree.search
------------------------------------------------------ -----------------------------
content-length  35'297 ps/iter    (33'693 ps … 337 ns) 34'613 ps 40'718 ps    109 ns
CONTENT-LENGTH  35'353 ps/iter    (34'034 ps … 368 ns) 34'954 ps 40'991 ps 82'356 ps
Content-Length  34'016 ps/iter (32'977 ps … 71'478 ps) 33'795 ps 39'354 ps 60'906 ps

summary for tree.search
  Content-Length
   1.04x faster than content-length
   1.04x faster than CONTENT-LENGTH
```

after:

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/TernarySearchTree.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark           time (avg)             (min … max)       p75       p99      p999
------------------------------------------------------ -----------------------------
• tree.search
------------------------------------------------------ -----------------------------
content-length  32'567 ps/iter    (30'964 ps … 379 ns) 31'919 ps 37'205 ps 93'200 ps
CONTENT-LENGTH  35'016 ps/iter    (33'761 ps … 174 ns) 34'716 ps 39'183 ps 89'211 ps
Content-Length  32'719 ps/iter (31'578 ps … 77'070 ps) 32'499 ps 36'625 ps 64'691 ps

summary for tree.search
  content-length
   1x faster than Content-Length
   1.08x faster than CONTENT-LENGTH
```